Removed useless classes and methods, code is optimized

+++ ClientForLaba/ClientForLaba/CipherUtils/PublicKey.cs
@@ -1,14 +1,15 @@
 ﻿using System;
+using System.Numerics;
 
 namespace ServerForLab.RsaCipher
 {
     [Serializable]
     public class PublicKey
     {
-        public long E { get; }
-        public long N { get; }
+        public BigInteger E { get; }
+        public BigInteger N { get; }
 
-        public PublicKey(long e, long n)
+        public PublicKey(BigInteger e, BigInteger n)
         {
             this.E = e;
             this.N = n;
